// shared/enum/permission.enum.ts
export enum PERMISSION_TYPE {
  CAN_CREATE_USER = "CAN_CREATE_USER",
  CAN_EDIT_USER = "CAN_EDIT_USER",
  CAN_ENABLE_USER = "CAN_ENABLE_USER",
  CAN_DELETE_USER = "CAN_DELETE_USER",
  CAN_CREATE_ROLE = "CAN_CREATE_ROLE",
  CAN_EDIT_ROLE = "CAN_EDIT_ROLE",
  CAN_DELETE_ROLE = "CAN_DELETE_ROLE",
  CAN_VIEW_REPORTS = "CAN_VIEW_REPORTS",
  CAN_CREATE_MONITOR = "CAN_CREATE_MONITOR",
  CAN_EDIT_MONITOR = "CAN_EDIT_MONITOR",
  CAN_ENABLE_MONITOR = "CAN_ENABLE_MONITOR",
  CAN_DELETE_MONITOR = "CAN_DELETE_MONITOR",
  CAN_GENERATE_NEW_URL = "CAN_GENERATE_NEW_URL",
}